Add spaces in coords and limits test
