+++ tests/test_dimensions.py
@@ -60,7 +60,7 @@ def test_check_n_obs():
 
 
 def test_combine_spaces():
-    combined_space = combine_spaces(spaces=[space1d_1, space1d_12])
+    combined_space = combine_spaces(space1d_1, space1d_12)
     combined_space2 = space1d_1 * space1d_12
     assert combined_space == space2d_2
     assert combined_space2 == space2d_2
