+++ zfit/core/space_new.py
@@ -15,6 +15,7 @@ from tensorflow.python.util.deprecation import deprecated
 
 import zfit
 from .baseobject import BaseObject
+from .coordinates import Coordinates, _convert_obs_to_str, _reorder_indices, convert_to_axes, convert_to_obs_str
 from .dimension import common_obs, common_axes, limits_overlap
 from .interfaces import ZfitLimit, ZfitOrderableDimensional, ZfitSpace
 from .. import z
@@ -154,249 +155,6 @@ def _sanitize_x_input(x, n_obs):
     if tf.get_static_value(x.shape[-1]) != n_obs:
         raise ShapeIncompatibleError("n_obs and the last dim of x do not agree. Assuming x has shape (..., n_obs)")
     return x
-
-
-def convert_to_numpy(obj):
-    if isinstance(obj, (tf.Tensor, tf.Variable)):
-        return obj.numpy()
-    else:
-        return np.array(obj)
-
-
-class Coordinates(ZfitOrderableDimensional):
-    def __init__(self, obs=None, axes=None):
-        obs, axes, n_obs = self._check_convert_obs_axes(obs, axes)
-        self._obs = obs
-        self._axes = axes
-        self._n_obs = n_obs
-
-    @staticmethod
-    def _check_convert_obs_axes(obs, axes):
-        if isinstance(obs, ZfitOrderableDimensional):
-            if axes is not None:
-                raise OverdefinedError("Cannot use (currently, please open an issue if desired) a"
-                                       " ZfitOrderableDimensional as obs with axes not None")
-            coord = obs
-            return coord.obs, coord.axes, coord.n_obs
-        obs = convert_to_obs_str(obs, container=tuple)
-        axes = convert_to_axes(axes, container=tuple)
-        if obs is None:
-            if axes is None:
-                raise CoordinatesUnderdefinedError("Neither obs nor axes specified")
-            else:
-                if any(not isinstance(ax, int) for ax in axes):
-                    raise ValueError(f"Axes have to be int. Currently: {axes}")
-            n_obs = len(axes)
-        else:
-            if any(not isinstance(ob, str) for ob in obs):
-                raise ValueError(f"Observables have to be strings. Currently: {obs}")
-            n_obs = len(obs)
-            if axes is not None:
-                if not len(obs) == len(axes):
-                    raise CoordinatesIncompatibleError("obs and axes do not have the same length.")
-        return obs, axes, n_obs
-
-    @property
-    def obs(self):
-        return self._obs
-
-    @property
-    def axes(self):
-        return self._axes
-
-    @property
-    def n_obs(self):
-        return self._n_obs
-
-    def with_obs(self, obs: Optional[ztyping.ObsTypeInput], allow_superset: bool = False, allow_subset: bool = False):
-        if obs is None:  # drop obs, check if there are axes
-            if self.axes is None:
-                raise AxesIncompatibleError("cannot remove obs (using None) for a Space without axes")
-            new_coords = type(self)(obs=obs, axes=self.axes)
-        else:
-            obs = _convert_obs_to_str(obs)
-
-            if not frozenset(obs) == frozenset(self.obs):
-
-                if not allow_superset and frozenset(obs).issuperset(self.obs):
-                    raise ObsIncompatibleError(
-                        f"Obs {obs} are a superset of {self.obs}, not allowed according to flag.")
-
-                if not allow_subset and frozenset(obs).issubset(self.obs):
-                    raise ObsIncompatibleError(
-                        f"Obs {obs} are a subset of {self.obs}, not allowed according to flag.")
-            new_indices = self.get_reorder_indices(obs=obs)
-            new_obs = self._reorder_obs(indices=new_indices)
-            new_axes = self._reorder_axes(indices=new_indices)
-            new_coords = type(self)(obs=new_obs, axes=new_axes)
-        return new_coords
-
-    def with_axes(self, axes: Optional[ztyping.AxesTypeInput], allow_superset: bool = False,
-                  allow_subset: bool = False) -> "zfit.Space":
-        """Sort by `axes` and return the new instance. `None` drops the axes.
-
-        Args:
-            axes ():
-            allow_superset (bool): Allow `axes` to be a superset of the `Spaces` axes
-
-        Returns:
-            :py:class:`~zfit.Space`
-        """
-        if axes is None:  # drop axes
-            if self.obs is None:
-                raise ObsIncompatibleError("Cannot remove axes (using None) for a Space without obs")
-            new_coords = type(self)(obs=self.obs, axes=axes)
-        else:
-            axes = _convert_axes_to_int(axes)
-            if not frozenset(axes) == frozenset(self.axes):
-                if not allow_superset and frozenset(axes).issuperset(self.axes):
-                    raise AxesIncompatibleError(
-                        f"Axes {axes} are a superset of {self.axes}, not allowed according to flag.")
-
-                if not allow_subset and frozenset(axes).issubset(self.axes):
-                    raise AxesIncompatibleError(
-                        f"Axes {axes} are a subset of {self.axes}, not allowed according to flag.")
-            new_indices = self.get_reorder_indices(axes=axes)
-            new_obs = self._reorder_obs(indices=new_indices)
-            new_axes = self._reorder_axes(indices=new_indices)
-            new_coords = type(self)(obs=new_obs, axes=new_axes)
-        return new_coords
-
-    def with_autofill_axes(self, overwrite: bool = False) -> "zfit.Space":
-        """Return a :py:class:`~zfit.Space` with filled axes corresponding to range(len(n_obs)).
-
-        Args:
-            overwrite (bool): If `self.axes` is not None, replace the axes with the autofilled ones.
-                If axes is already set, don't do anything if `overwrite` is False.
-
-        Returns:
-            :py:class:`~zfit.Space`
-        """
-        if self.axes and not overwrite:
-            raise ValueError("overwrite is not allowed but axes are already set.")
-        new_coords = type(self)(obs=self.obs, axes=range(self.n_obs))
-        return new_coords
-
-    def _reorder_obs(self, indices: Tuple[int]) -> ztyping.ObsTypeReturn:
-        obs = self.obs
-        if obs is not None:
-            obs = tuple(obs[i] for i in indices)
-        return obs
-
-    def _reorder_axes(self, indices: Tuple[int]) -> ztyping.AxesTypeReturn:
-        axes = self.axes
-        if axes is not None:
-            axes = tuple(axes[i] for i in indices)
-        return axes
-
-    def get_reorder_indices(self, obs: ztyping.ObsTypeInput = None,
-                            axes: ztyping.AxesTypeInput = None) -> Tuple[int]:
-        """Indices that would order `self.obs` as `obs` respectively `self.axes` as `axes`.
-
-        Args:
-            obs ():
-            axes ():
-
-        Returns:
-
-        """
-        obs_none = obs is None
-        axes_none = axes is None
-
-        obs_is_defined = self.obs is not None and not obs_none
-        axes_is_defined = self.axes is not None and not axes_none
-        if not (obs_is_defined or axes_is_defined):
-            raise ValueError(
-                "Neither the `obs` (argument and on instance) nor `axes` (argument and on instance) are defined.")
-
-        if obs_is_defined:
-            old, new = self.obs, [o for o in obs if o in self.obs]
-        else:
-            old, new = self.axes, [a for a in axes if a in self.axes]
-
-        new_indices = _reorder_indices(old=old, new=new)
-        return new_indices
-
-    def reorder_x(self, x: Union[tf.Tensor, np.ndarray], *, x_obs: ztyping.ObsTypeInput = None,
-                  x_axes: ztyping.AxesTypeInput = None, func_obs: ztyping.ObsTypeInput = None,
-                  func_axes: ztyping.AxesTypeInput = None) -> Union[tf.Tensor, np.ndarray]:
-        """Reorder x in the last dimension either according to its own obs or assuming a function ordered with func_obs.
-
-        There are two obs or axes around: the one associated with this Coordinate object and the one associated with x.
-        If x_obs or x_axes is given, then this is assumed to be the obs resp. the axes of x and x will be reordered
-        according to `self.obs` resp. `self.axes`.
-
-        If func_obs resp. func_axes is given, then x is assumed to have `self.obs` resp. `self.axes` and will be
-        reordered to align with a function ordered with `func_obs` resp. `func_axes`.
-
-        Switching `func_obs` for `x_obs` resp. `func_axes` for `x_axes` inverts the reordering of x.
-
-        Args:
-            x (tensor-like): Tensor to be reordered, last dimension should be n_obs resp. n_axes
-            x_obs: Observables associated with x. If both, x_obs and x_axes are given, this has precedency over the
-                latter.
-            x_axes: Axes associated with x.
-            func_obs: Observables associated with a function that x will be given to. Reorders x accordingly and assumes
-                self.obs to be the obs of x. If both, `func_obs` and `func_axes` are given, this has precedency over the
-                latter.
-            func_axes: Axe associated with a function that x will be given to. Reorders x accordingly and assumes
-                self.axes to be the axes of x.
-
-        Returns:
-
-        """
-        x_reorder = x_obs is not None or x_axes is not None
-        func_reorder = func_obs is not None or func_axes is not None
-        if not (x_reorder ^ func_reorder):
-            raise ValueError("Either specify `x_obs/axes` or `func_obs/axes`, not both.")
-        obs_defined = x_obs is not None or func_obs is not None
-        axes_defined = x_axes is not None or func_axes is not None
-        if obs_defined and self.obs:
-            if x_reorder:
-                coord_old = x_obs
-                coord_new = self.obs
-            elif func_reorder:
-                coord_new = func_obs
-                coord_old = self.obs
-            else:
-                assert False, 'bug, should never be reached'
-
-        elif axes_defined and self.axes:
-            if x_reorder:
-                coord_old = x_axes
-                coord_new = self.axes
-            elif func_reorder:
-                coord_new = func_axes
-                coord_old = self.axes
-            else:
-                assert False, 'bug, should never be reached'
-        else:
-            raise ValueError("Obs and self.obs or axes and self. axes not properly defined. Can only reorder on defined"
-                             " coordinates.")
-
-        new_indices = _reorder_indices(old=coord_old, new=coord_new)
-
-        x = z.unstable.gather(x, indices=new_indices, axis=-1)
-        return x
-
-    def __eq__(self, other):
-        if not isinstance(other, Coordinates):
-            return NotImplemented
-        obs_equal = False
-        axes_equal = False
-        if self.obs is not None and other.obs is not None:
-            obs_equal = frozenset(self.obs) == frozenset(other.obs)
-
-        if self.axes is not None and other.axes is not None:
-            axes_equal = frozenset(self.axes) == frozenset(other.axes)
-        equal = obs_equal or axes_equal
-        return equal
-
-    def __hash__(self):
-        return 42  # always check with equal...  maybe change in future, use dict that checks for different things.
-
-    def __repr__(self):
-        return f"<zfit Coordinates obs={self.obs}, axes={self.axes}"
 
 
 def is_range_definition(limit):
@@ -801,10 +559,7 @@ class BaseSpace(ZfitSpace, BaseObject):
     def equal(self, other, allow_graph):
         if not isinstance(other, ZfitSpace):
             return NotImplemented
-        return equal_space(other, allow_graph=allow_graph)
-        # limits_equal = self.rect_limits == other.rect_limits  # TODO: improve! What about 'inside'?
-        # coords_equal = self.coords == other.coords
-        # return z.unstable.logical_and(limits_equal, coords_equal)
+        return equal_space(self, other, allow_graph=allow_graph)
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, ZfitSpace):
@@ -863,6 +618,9 @@ class BaseSpace(ZfitSpace, BaseObject):
     def reorder_x(self, x, x_obs, x_axes, func_obs, func_axes):
         return self.coords.reorder_x(x, x_obs=x_obs, x_axes=x_axes,
                                      func_obs=func_obs, func_axes=func_axes)
+
+    def __len__(self):
+        return sum(1 for _ in self)
 
 
 class Space(BaseSpace):
@@ -996,7 +754,7 @@ class Space(BaseSpace):
             else:
                 obs_limit_dict = {}
                 for axes_lim, lim in input_limits['axes'].items():
-                    obs_coords = tuple(obs[axes.index[ax]] for ax in axes_lim)
+                    obs_coords = tuple(obs[axes.index(ax)] for ax in axes_lim)
                     obs_limit_dict[obs_coords] = lim
             limits_dict['obs'] = obs_limit_dict
 
@@ -1006,7 +764,7 @@ class Space(BaseSpace):
             else:
                 axes_limit_dict = {}
                 for obs_lim, lim in input_limits['obs'].items():
-                    axes_coords = tuple(axes[obs.index[ob]] for ob in obs_lim)
+                    axes_coords = tuple(axes[obs.index(ob)] for ob in obs_lim)
                     axes_limit_dict[axes_coords] = lim
             limits_dict['axes'] = axes_limit_dict
 
@@ -1302,7 +1060,7 @@ class Space(BaseSpace):
                 return self
             if self.axes is None:
                 raise AxesIncompatibleError("cannot remove obs (using None) for a Space without axes")
-            new_limits = self._limits_dict['obs']
+            new_limits = self._limits_dict
             new_space = self.copy(obs=obs, limits=new_limits)
         else:
             obs = _convert_obs_to_str(obs)
@@ -1327,7 +1085,7 @@ class Space(BaseSpace):
                 return self
             if self.obs is None:
                 raise ObsIncompatibleError("Cannot remove axes (using None) for a Space without obs")
-            new_limits = self._limits_dict['axes']
+            new_limits = self._limits_dict.copy()
             new_space = self.copy(axes=axes, limits=new_limits)
         else:
             axes = convert_to_axes(axes)
@@ -1335,9 +1093,11 @@ class Space(BaseSpace):
                 if not len(axes) == len(self.obs):
                     raise AxesIncompatibleError(f"Trying to set axes {axes} to object with obs {self.obs}")
                 new_limits = self._limits_dict['obs'].copy()
-                for obs, limit in self._limits_dict['obs']:
-                    ax = tuple(axes[obs.index[ob]] for ob in obs)
+                new_limits['axes'] = {}
+                for obs, limit in self._limits_dict['obs'].items():
+                    ax = tuple(axes[obs.index(ob)] for ob in obs)
                     new_limits['axes'][ax] = limit
+                new_space = self.copy(axes=axes, limits=new_limits)
             else:
 
                 coords = self.coords.with_axes(axes=axes, allow_superset=allow_superset, allow_subset=allow_subset)
@@ -1403,7 +1163,8 @@ class Space(BaseSpace):
                     raise BehaviorUnderDiscussion(f"The axes of self {self.axes} and {coords_obs.axes} do not agree.")
                 coords_obs = Coordinates(obs=self.obs, axes=self.axes or coords_obs.axes)
             new_space_obs = self.with_obs(coords.obs, allow_superset=allow_superset, allow_subset=allow_subset)
-            new_space_obs.coords = coords  # are the same
+            # new_space_obs.coords._axes = coords_obs.axes
+            # new_space_obs = new_space_obs.with_axes(coords_obs.axes) # are the same
 
         if self.axes is not None and coords.axes is not None:
             coords_axes = coords.with_axes(self.axes)
@@ -1412,7 +1173,8 @@ class Space(BaseSpace):
                     raise BehaviorUnderDiscussion(f"The obs of self {self.obs} and {coords_axes.obs} do not agree.")
                 coords_axes = Coordinates(obs=self.obs or coords_axes.obs, axes=self.axes)
             new_space_axes = self.with_axes(coords.axes, allow_superset=allow_superset, allow_subset=allow_subset)
-            new_space_axes.coords = coords_axes
+            # new_space_axes.coords._obs = coords_axes.obs
+            # new_space_axes = new_space_axes.with_obs(coords_axes.obs)
 
         if new_space_obs is not None and new_space_axes is not None:
             if new_space_obs.axes != new_space_axes.axes or new_space_obs.obs != new_space_axes.obs:
@@ -1444,9 +1206,10 @@ class Space(BaseSpace):
         Returns:
             :py:class:`~zfit.Space`
         """
+        new_coords = self.coords.with_autofill_axes(overwrite=overwrite)
+        # new_space = self.with_coords(new_coords)
         if self.axes is None or overwrite:
-            new_axes = tuple(range(self.n_obs))
-            new_space = self.copy(axes=new_axes)
+            new_space = self.copy(axes=new_coords.axes)
         else:
             new_space = self
 
@@ -1824,22 +1587,6 @@ def compare_limits_dict(dict1: Mapping, dict2: Mapping, comparator: Callable) ->
     return True
 
 
-def _convert_axes_to_int(axes):
-    if isinstance(axes, ZfitSpace):
-        axes = axes.axes
-    else:
-        axes = convert_to_container(axes, container=tuple)
-    return axes
-
-
-def _convert_obs_to_str(obs):
-    if isinstance(obs, ZfitSpace):
-        obs = obs.obs
-    else:
-        obs = convert_to_container(obs, container=tuple)
-    return obs
-
-
 def flatten_spaces(spaces):
     return tuple(s for space in spaces for s in space)
 
@@ -2128,11 +1875,6 @@ def convert_to_space(obs: Optional[ztyping.ObsTypeInput] = None, axes: Optional[
     return space
 
 
-def _reorder_indices(old: Union[List, Tuple], new: Union[List, Tuple]) -> Tuple[int]:
-    new_indices = tuple(old.index(o) for o in new)
-    return new_indices
-
-
 def no_norm_range(func):
     """Decorator: Catch the 'norm_range' kwargs. If not None, raise NormRangeNotImplementedError."""
     parameters = inspect.signature(func).parameters
@@ -2215,42 +1957,6 @@ def supports(*, norm_range: bool = False, multiple_limits: bool = False) -> Call
     return create_deco_stack
 
 
-def convert_to_axes(axes, container=tuple):
-    """Convert `obs` to the list of obs, also if it is a :py:class:`~ZfitSpace`. Return None if axes is None.
-
-    """
-    if axes is None:
-        return axes
-    axes = convert_to_container(value=axes, container=container)
-    new_axes = []
-    for axis in axes:
-        if isinstance(axis, ZfitSpace):
-            if len(axis) > 1:
-                raise WorkInProgressError("Not implemented, uniqueify?")
-            new_axes.extend(axis.obs)
-        else:
-            new_axes.append(axis)
-    return container(new_axes)
-
-
-def convert_to_obs_str(obs, container=tuple):
-    """Convert `obs` to the list of obs, also if it is a :py:class:`~ZfitSpace`. Return None if obs is None.
-
-    """
-    if obs is None:
-        return obs
-    obs = convert_to_container(value=obs, container=container)
-    new_obs = []
-    for ob in obs:
-        if isinstance(ob, ZfitSpace):
-            if len(ob) > 1:
-                raise WorkInProgressError("Not implemented, uniqueify?")
-            new_obs.extend(ob.obs)
-        else:
-            new_obs.append(ob)
-    return container(new_obs)
-
-
 def contains_tensor(object):
     tensor_found = tf.is_tensor(object)
     with suppress(TypeError):
@@ -2286,7 +1992,7 @@ def limits_consistent(spaces: Iterable["zfit.Space"]):
         bool:
     """
     try:
-        new_space = combine_spaces(spaces=spaces)
+        new_space = combine_spaces(*spaces)
     except LimitsIncompatibleError:
         return False
     return bool(new_space)
